:art: UPDATE Picture Module invoke RESTful API (POST & PUT)

+++ src/main/java/com/ssm/promotion/core/admin/PictureController.java
@@ -6,11 +6,15 @@ import com.ssm.promotion.core.entity.PageBean;
 import com.ssm.promotion.core.entity.Picture;
 import com.ssm.promotion.core.service.PictureService;
 import com.ssm.promotion.core.util.DateUtil;
+import com.ssm.promotion.core.util.ResponseUtil;
 import org.apache.log4j.Logger;
+import net.sf.json.JSONArray;
+import net.sf.json.JSONObject;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpServletResponse;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -22,6 +26,44 @@ public class PictureController {
     private PictureService pictureService;
     private static final long serialVersionUID = 1L;
     private static final Logger log = Logger.getLogger(PictureController.class);// 日志文件
+
+    /**
+     * 查找相应的数据集合
+     *
+     * @param page
+     * @param rows
+     * @param picture
+     * @param response
+     * @return
+     * @throws Exception
+     */
+    @RequestMapping("/datagrid")
+    public String list(
+            @RequestParam(value = "page", required = false) String page,
+            @RequestParam(value = "rows", required = false) String rows,
+            Picture picture, HttpServletResponse response) throws Exception {
+        Map<String, Object> map = new HashMap<String, Object>();
+        if (page != null && rows != null) {
+            PageBean pageBean = new PageBean(Integer.parseInt(page),
+                    Integer.parseInt(rows));
+            map.put("start", pageBean.getStart());
+            map.put("size", pageBean.getPageSize());
+        }
+        if (picture != null) {
+            map.put("id", picture.getId() + "");
+            map.put("type", picture.getType() + "");
+            map.put("grade", picture.getGrade() + "");
+        }
+        List<Picture> pictureList = pictureService.findPicture(map);
+        Long total = pictureService.getTotalPicture(map);
+        JSONObject result = new JSONObject();
+        JSONArray jsonArray = JSONArray.fromObject(pictureList);
+        result.put("rows", jsonArray);
+        result.put("total", total);
+        log.info("request: pictures/datagrid , map: " + map.toString());
+        ResponseUtil.write(response, result);
+        return null;
+    }
 
     /**
      * 查找相应的数据集合
@@ -70,7 +112,7 @@ public class PictureController {
      */
     @RequestMapping(value = "/", method = RequestMethod.POST)
     @ResponseBody
-    public Result save(Picture picture)
+    public Result save(@RequestBody Picture picture)
             throws Exception {
         int resultTotal = 0;
 
@@ -96,9 +138,11 @@ public class PictureController {
      */
     @RequestMapping(value = "/", method = RequestMethod.PUT)
     @ResponseBody
-    public Result update(Picture picture)
+    public Result update(@RequestBody Picture picture)
             throws Exception {
         int resultTotal = 0;
+        
+        picture.setTime(DateUtil.getCurrentDateStr());
 
         resultTotal = pictureService.updatePicture(picture);
 
@@ -122,6 +166,9 @@ public class PictureController {
     @RequestMapping(value = "/{ids}", method = RequestMethod.DELETE)
     public Result delete(@PathVariable(value = "ids") String ids
     ) throws Exception {
+        if (ids.length() > 20) {
+            return ResultGenerator.genFailResult("ERROR");
+        }
         String[] idsStr = ids.split(",");
         for (int i = 0; i < idsStr.length; i++) {
             pictureService.deletePicture(idsStr[i]);
